Some compression stats

+++ projects/pandora/include/pandora/svo/sparse_voxel_dag.h
@@ -28,7 +28,7 @@ public:
 	size_t size() const { return m_allocator.size() * sizeof(decltype(m_allocator)::value_type); }
 
 private:
-	using RelativeNodeOffset = uint16_t; // Either uint32_t or uint16_t
+	using RelativeNodeOffset = uint32_t; // Either uint32_t or uint16_t
 	using AbsoluteNodeOffset = size_t;
 
     // NOTE: child pointers are stored directly after the descriptor
